🐛 Fix saving link patterns
