+++ package.json
@@ -18,6 +18,7 @@
     "@emotion/styled": "^11.10.6",
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.13.2",
+    "@radix-ui/react-hover-card": "^1.0.6",
     "@radix-ui/react-slot": "^1.0.2",
     "@reduxjs/toolkit": "^1.9.3",
     "class-variance-authority": "^0.6.0",
