+++ src/app/ui/Support.tsx
@@ -1,0 +1,8 @@
+import { HelpCircle } from "lucide-react";
+
+export const Support = () => (
+  <div className="flex items-center text-sm font-medium text-gray-400">
+    <HelpCircle className="mr-1 h-4 w-4 stroke-[2px]" />
+    Support
+  </div>
+);
